pyproject.toml
[tool.pytest.ini_options]
log_cli = 1
log_level = "INFO"
addopts = "-s -v --clean-alluredir --alluredir=allure-results --allure-no-capture"
markers = [
    "get: for test with GET request to API",
    "post: for test with POST request to API",
    "put: for test with PUT request to API",
    "patch: for test with PATCH request to API",
    "delete: for test with DELETE request to API"
]

[tool.poetry]
name = "qaguru-18-dkudriavtseva"
version = "0.1.0"
description = ""
authors = ["Daria Kudriavtseva <situation96@mail.ru>"]
readme = "README.md"
packages = [{include = "qaguru_18_dkudriavtseva"}]

[tool.poetry.dependencies]
python = "^3.11"
pytest = "7.3.1"
requests = "^2.31.0"
voluptuous = "^0.13.1"
pytest-voluptuous = "^1.2.0"
allure-pytest = "^2.13.2"
curlify = "^2.2.1"
selene = "2.0.0rc2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
